requirements.txt
asgiref==3.9.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
dj-database-url==0.5.0
Django==5.2.5
django-autoslug==1.9.9
django-cities-light==3.10.1
django-countries==7.6.1
django-crontab==0.7.1
django-debug-toolbar==6.0.0
django-silk==5.4.0
djangorestframework==3.15.1
exceptiongroup==1.2.2
geographiclib==2.0
geopy==2.4.1
gprof2dot==2025.4.14
gunicorn==23.0.0
h11==0.14.0
idna==3.10
importlib-metadata==6.7.0
newsapi-python==0.2.7
outcome==1.3.0.post0
packaging==24.0
Pillow==9.5.0
progressbar2==4.2.0
psycopg2-binary==2.9.10
pymemcache==4.0.0
PySocks==1.7.1
python-decouple==3.8
python-utils==3.5.2
pytz==2024.2
pyuca==1.2
requests==2.31.0
selenium==4.11.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
Unidecode==1.3.8
urllib3==2.0.7
uvicorn==0.22.0
whitenoise==5.3.0
wsproto==1.2.0
zipp==3.15.0
psycopg2
